Bug fix#2
