+++ index.html
@@ -251,7 +251,7 @@
   </template>
 
   <!-- Скрипты -->
-  <script src="./js/functions.js"></script>
+  <script src="./js/main.js" type="module"></script>
 
 </body>
